<refact>: uploadpost/editpost 페이지 중복코드 추출
